chore: export as `Spawn`
This is a breaking change and is intended to help distinguish big Spawn
from little spawn (constructor vs. instance). While the main thread has
explicit access to the `Spawn` constructor, workers only have access to
an instance of Spawn. Also.. the fact that a workers default Spawn
instance is named `spawn` does not help.. I hope to address this in the
future.